Parameter to install from online or local scripts

+++ setup.ps1
@@ -1,25 +1,51 @@
-Param([switch]$InstallOptional)
+Param (
+    [Parameter(Position=0)]
+    [switch]
+    $InstallOptional,
+
+    [Parameter()]
+    [switch]
+    $Online
+)
 
 function Test-Admin {
     $currentUser = New-Object Security.Principal.WindowsPrincipal $([Security.Principal.WindowsIdentity]::GetCurrent())
     $currentUser.IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator)
 }
 
+$currentFolder = $PWD
+
 if ((Test-Admin) -eq $false)  {
-    Start-Process PowerShell.exe -WorkingDirectory $PWD -Verb RunAs -ArgumentList ('-noprofile -noexit -file "{0}" -elevated' -f ($myinvocation.MyCommand.Definition))        
+    $cmd = $myinvocation.MyCommand.Path 
+    if ($InstallOptional) {
+        $cmd += " -InstallOptional"
+    } 
+    if ($Online) {
+        $cmd += " -Online"
+    }
+
+    Start-Process PowerShell.exe -WorkingDirectory $PWD -Verb RunAs -ArgumentList `
+        ('-noprofile -noexit -Command ' + $cmd)
     exit
 }
 
-Write-Host -ForegroundColor Green "Downloading all required scripts..."
-$currentFolder = $($PWD)
-$scriptsPath = $env:TEMP + "\dev-strap"
-$prerequisitesPath = $scriptsPath+ "\dev-strap-main\pre-requisites"
-New-Item -Path $scriptsPath -ItemType Directory | Out-Null
-Set-Location $scriptsPath
-Invoke-WebRequest -uri https://github.com/slobodnic/dev-strap/archive/refs/heads/main.zip -OutFile scripts.zip
-Expand-Archive scripts.zip -DestinationPath .
+$scriptsPath = Split-Path -parent $myinvocation.MyCommand.Path
+$prerequisitesPath = Join-Path $scriptsPath -ChildPath pre-requisites
 
-'running with full privileges'
+if ($Online) {
+    Write-Host -ForegroundColor Green "Downloading all required scripts..."
+    $scriptsPath = Join-Path $env:TEMP -ChildPath dev-strap
+    New-Item -Path $scriptsPath -ItemType Directory | Out-Null
+    Set-Location $scriptsPath
+    $pwd
+    Invoke-WebRequest -uri https://github.com/slobodnic/dev-strap/archive/refs/heads/main.zip -OutFile scripts.zip
+    Expand-Archive scripts.zip -DestinationPath $scriptsPath
+    $scriptsPath = Join-Path $scriptsPath -ChildPath dev-strap-main
+    $prerequisitesPath = Join-Path $scriptsPath -ChildPath pre-requisites
+} else {
+    Write-Host -ForegroundColor Green "Installing from local scripts..."
+}
+
 Write-Host -ForegroundColor Green "Installing pre-requisites" 
 # The pre-requisit scripts are in order in which they need to install
 
@@ -37,21 +63,23 @@ refreshenv;
 
 Write-Host -ForegroundColor Green "Installing the tools" 
 
-Set-Location $scriptsPath'\dev-strap-main\scripts'
-Get-ChildItem -Filter '*.ps1' $scriptsPath'\dev-strap-main\scripts' | ForEach-Object {
+Set-Location $scriptsPath'\scripts'
+Get-ChildItem -Filter '*.ps1' $scriptsPath'\scripts' | ForEach-Object {
     & $_.FullName
 }
 
 if ($InstallOptional)
 {
     Write-Host -ForegroundColor Green "Installing optional tools" 
-    Set-Location $scriptsPath'\dev-strap-main\optional'
-    Get-ChildItem -Filter '*.ps1' $scriptsPath'\dev-strap-main\optional' | ForEach-Object {
+    Set-Location $scriptsPath'\optional'
+    Get-ChildItem -Filter '*.ps1' $scriptsPath'\optional' | ForEach-Object {
         & $_.FullName
     }
 }
 
 Set-Location $currentFolder
-Remove-Item $env:TEMP\dev-strap\ -Force -Recurse
+if ($Online) {
+    Remove-Item $env:TEMP\dev-strap\ -Force -Recurse
+}
 # stop-process -Id $PID
 # logoff.exe
